publish result set
